Use the settings tab title for the window title (#5555)
This commit will change PreferenceWindowController.loadTab to set the
window's title to be the title of the selected tab.

This change is required to make IINA adhere to Apple's Human Interface
Guidelines which require the tab's title be used. Although IINA hides
the window's title it still can be seen in the Window menu and in the
dock menu.