Fix no longer remembers last used display, #5680 (#5681)
* Fix no longer remembers last used display, #5680

This commit will change MainWindowController.showWindow to detect when
NSWindowController.showWindow changes the window's origin and move the
window back to the previously used display.

* Fix no longer remembers last used display, #5680

This commit will change MainWindowController.showWindow to detect when
NSWindowController.showWindow changes the window's origin and move the
window back to the previously used display.

* Fix no longer remembers last used display, #5680

This commit will:
- Add log messages when the workaround is being applied
- Add comment indicating the AppKit problem was fixed in macOS Tahoe

* Fix no longer remembers last used display, #5680

This commit will:
- Change MainWindowController.showWindow to detect when
  NSWindowController.showWindow changes the window's origin and move the
  window back to the previously used display.
- Add an internal enableWrongScreenWorkaround setting that can be used
  to disable the workaround should a problem be found with it
- Add additional logging to determineScreenToUse and
  handleVideoSizeChange in  MainWindowController

This works around an AppKit defect where when a window is first
displayed NSWindowController.showWindow may move the window to a
different screen. This AppKit defect has been fixed as of macOS Tahoe.